split scaffold test harness helpers
